+++ src/test/java/ru/progwards/tasktracker/service/facade/impl/worklog/WorkLogCreateServiceTest.java
@@ -14,6 +14,7 @@ import java.time.ZonedDateTime;
 import static org.hamcrest.CoreMatchers.equalTo;
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.junit.jupiter.api.Assertions.assertNotNull;
+import static org.junit.jupiter.api.Assertions.fail;
 
 /**
  * тестирование сервиса создания лога
@@ -49,12 +50,15 @@ class WorkLogCreateServiceTest {
                 .map(WorkLog::getId)
                 .orElse(null);
 
-        WorkLog workLog = getService.get(id);
+        if (id != null) {
+            WorkLog workLog = getService.get(id);
 
-        assertNotNull(workLog);
+            assertNotNull(workLog);
 
-        assertThat(workLog.getDescription(), equalTo("Description Log CreateService"));
+            assertThat(workLog.getDescription(), equalTo("Description Log CreateService"));
 
-        removeService.remove(workLog);
+            removeService.remove(workLog);
+        } else
+            fail();
     }
 }
